veyron2/vdl: Generate named types for Java, with basic setters/getters.
Will implement Map/Set/List interfaces in subsequent CL.

+++ src/com/veyron/runtimes/google/vsync/VeyronConsts.java
@@ -6,7 +6,7 @@ package com.veyron.runtimes.google.vsync;
 public class VeyronConsts { 
 	/* The following constants originate in file: vsync.vdl. */
 	// NodeRec type log record adds a new node in the dag.
-public static final byte NODE_REC = 0;
+	public static final byte NODE_REC = (byte)0;
 	// LinkRec type log record adds a new link in the dag.
-public static final byte LINK_REC = 1;
+	public static final byte LINK_REC = (byte)1;
 }
